requirements.txt
fastapi
uvicorn
pdfplumber
spacy
openai
streamlit
python-dotenv
python-multipart
sentence-transformers
scikit-learn
